requirements.azure.txt
opencensus-ext-azure==1.0.7
